+++ src/img/test/basics.test.js
@@ -20,10 +20,10 @@
  *
  * CDDL HEADER END
  *
- * Copyright (c) 2013, Joyent, Inc. All rights reserved.
+ * Copyright (c) 2014, Joyent, Inc. All rights reserved.
  *
  * * *
- * Some base imgadm tests.
+ * Some basic imgadm tests.
  */
 
 var p = console.log;
@@ -152,6 +152,24 @@ test('`imgadm info BOGUS_UUID` ImageNotInstalled error, rv 3', function (t) {
             'ImageNotInstalled error code on stderr');
         t.equal(stdout, '', 'no stdout');
 
+        t.end();
+    });
+});
+
+
+// OS-2981: we broke 'imgadm avail' with a DSAPI source once.
+test('imgadm sources -a DSAPI', function (t) {
+    exec('imgadm sources -a https://datasets.joyent.com/datasets',
+            function (err, stdout, stderr) {
+        t.ifError(err, err);
+        t.end();
+    });
+});
+test('imgadm avail', function (t) {
+    exec('imgadm avail -o source,uuid,name | grep datasets.joyent.com',
+            function (err, stdout, stderr) {
+        t.ifError(err, err);
+        t.ok(/base/.test(stdout), 'datasets.joyent.com provides a "base" img');
         t.end();
     });
 });